Finished converting MenuModal from old way of creating modals to dialog HTML element
