+++ src/components/index.js
@@ -11,7 +11,7 @@ export { default as CardUtils } from "./CardUtils";
 export { default as LetterTile } from "./LetterTile";
 export { default as SyllableTile } from "./SyllableTile";
 export { default as FactorUnitAnimations } from "./FactorUnitAnimations";
-export { default as Menu } from "./Menu";
+export { default as MenuModal } from "./MenuModal";
 export { default as Header } from "./Header";
 export { default as LikeButton } from "./LikeButton";
 export { default as SlideShow } from "./SlideShow";
